[ADD] sale_product_pack: sale product packs
