+++ oca_dependencies.txt
@@ -2,4 +2,5 @@ account-invoicing
 partner-contact
 server-tools
 web
+product-attribute https://github.com/Tecnativa/product-attribute 12.0-add-product_pack
 bank-payment
