--- product_pack/README.rst
@@ -1,75 +1,0 @@
-.. |company| replace:: ADHOC SA
-
-.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
-   :alt: ADHOC SA
-   :target: https://www.adhoc.com.ar
-
-.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
-
-.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
-   :target: https://www.gnu.org/licenses/agpl
-   :alt: License: AGPL-3
-
-============
-Product Pack
-============
-
-This module add "Product Pack" funcionality to sale orders. For product you can add some other product to sell with them, in the sale order you choose the product and then automaticaly added the products that set in Product selected.
-
-Nota sobre permisos. Odoo por defecto deja que grupos "Manager" e "Inventario / User" puedan gestionar productos. Para no tener que agregar dependencia a stock ni hacer modulo puente con las diversas apps que crean esos grupos, damos permiso sobre "product.pack.line" a "Empleados" para que de esta manera, cualquiera que pueda gestionar productos pueda gestionar la info del pack.
-
-TODO: check modifications of this fork https://github.com/mohamedhagag/dvit-odoo/tree/10.0/product_pack
-
-Installation
-============
-
-To install this module, you need to:
-
-#. Just install this module.
-
-Configuration
-=============
-
-To configure this module, you need to:
-
-#. Go to product and set "Product Pack", then set the products has to include in the product pack of the product.
-#. Set type of product pack.
-
-Usage
-=====
-
-To use this module, you need to:
-
-#. Create an sale order and set product that has product pack choosed.
-
-.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-   :alt: Try me on Runbot
-   :target: http://runbot.adhoc.com.ar/
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues
-<https://github.com/ingadhoc/product/issues>`_. In case of trouble, please
-check there if your issue has already been reported. If you spotted it first,
-help us smashing it by providing a detailed and welcomed feedback.
-
-Credits
-=======
-
-Images
-------
-
-* |company| |icon|
-
-Contributors
-------------
-
-Maintainer
-----------
-
-|company_logo|
-
-This module is maintained by the |company|.
-
-To contribute to this module, please visit https://www.adhoc.com.ar.
