--- product_pack/i18n/en.po
@@ -1,207 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * product_pack
-# 
-# Translators:
-msgid ""
-msgstr ""
-"Project-Id-Version: odoo-addons (8.0)\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-09-09 23:57+0000\n"
-"PO-Revision-Date: 2015-09-10 00:00+0000\n"
-"Last-Translator: Juan Jose Scarafia <scarafia.juanjose@gmail.com>\n"
-"Language-Team: English (http://www.transifex.com/adhoc/ingadhoc-odoo-addons-8-0/language/en/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: en\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-
-#. module: product_pack
-#: help:product.template,pack_price_type:0
-msgid ""
-"\n"
-"        * Totalice Price: Sum individual prices on the product pack price.\n"
-"        * Fixed Price: Price of this product instead of components prrices.\n"
-"        * Components Price: Components prices plast pack price.\n"
-"        "
-msgstr "\n        * Totalice Price: Sum individual prices on the product pack price.\n        * Fixed Price: Price of this product instead of components prrices.\n        * Components Price: Components prices plast pack price.\n        "
-
-#. module: product_pack
-#: selection:product.template,pack_price_type:0
-msgid "Components Prices"
-msgstr "Components Prices"
-
-#. module: product_pack
-#: field:product.pack.line,create_uid:0
-#: field:sale.order.line.pack.line,create_uid:0
-msgid "Created by"
-msgstr "Created by"
-
-#. module: product_pack
-#: field:product.pack.line,create_date:0
-#: field:sale.order.line.pack.line,create_date:0
-msgid "Created on"
-msgstr "Created on"
-
-#. module: product_pack
-#: field:sale.order.line,pack_depth:0
-msgid "Depth"
-msgstr "Depth"
-
-#. module: product_pack
-#: help:sale.order.line,pack_depth:0
-msgid "Depth of the product if it is part of a pack."
-msgstr "Depth of the product if it is part of a pack."
-
-#. module: product_pack
-#: selection:product.template,pack_price_type:0
-msgid "Fixed Price"
-msgstr "Fixed Price"
-
-#. module: product_pack
-#: field:product.pack.line,id:0 field:sale.order.line.pack.line,id:0
-msgid "ID"
-msgstr "ID"
-
-#. module: product_pack
-#: field:product.pack.line,write_uid:0
-#: field:sale.order.line.pack.line,write_uid:0
-msgid "Last Updated by"
-msgstr "Last Updated by"
-
-#. module: product_pack
-#: field:product.pack.line,write_date:0
-#: field:sale.order.line.pack.line,write_date:0
-msgid "Last Updated on"
-msgstr "Last Updated on"
-
-#. module: product_pack
-#: field:sale.order.line,pack_child_line_ids:0
-msgid "Lines in pack"
-msgstr "Lines in pack"
-
-#. module: product_pack
-#: help:product.product,pack_line_ids:0
-#: help:product.product,used_pack_line_ids:0
-msgid "List of products that are part of this pack."
-msgstr "List of products that are part of this pack."
-
-#. module: product_pack
-#: model:res.groups,name:product_pack.group_pack
-msgid "Manage Packs On Sales Orders"
-msgstr "Manage Packs On Sales Orders"
-
-#. module: product_pack
-#: field:sale.order.line.pack.line,order_line_id:0
-msgid "Order Line"
-msgstr "Order Line"
-
-#. module: product_pack
-#: view:product.product:product_pack.nan_product_pack_form
-#: field:sale.order.line,pack_parent_line_id:0
-msgid "Pack"
-msgstr "Pack"
-
-#. module: product_pack
-#: field:sale.order.line,pack_line_ids:0
-msgid "Pack Lines"
-msgstr "Pack Lines"
-
-#. module: product_pack
-#: field:product.template,pack_price_type:0
-msgid "Pack Price Type"
-msgstr "Pack Price Type"
-
-#. module: product_pack
-#: view:product.pack.line:product_pack.product_pack_line_form
-msgid "Pack Product Line"
-msgstr "Pack Product Line"
-
-#. module: product_pack
-#: view:product.pack.line:product_pack.product_pack_line_tree
-msgid "Pack Product Lines"
-msgstr "Pack Product Lines"
-
-#. module: product_pack
-#: field:product.product,pack_line_ids:0
-#: field:product.product,used_pack_line_ids:0
-msgid "Pack Products"
-msgstr "Pack Products"
-
-#. module: product_pack
-#: code:addons/product_pack/pack.py:159 code:addons/product_pack/pack.py:164
-#, python-format
-msgid ""
-"Pack lines products company must be the same as the                    "
-"parent product company"
-msgstr "Pack lines products company must be the same as the                    parent product company"
-
-#. module: product_pack
-#: field:product.template,pack:0
-msgid "Pack?"
-msgstr "Pack?"
-
-#. module: product_pack
-#: field:product.pack.line,parent_product_id:0
-msgid "Parent Product"
-msgstr "Parent Product"
-
-#. module: product_pack
-#: model:ir.model,name:product_pack.model_product_product
-#: field:product.pack.line,product_id:0
-#: field:sale.order.line.pack.line,product_id:0
-msgid "Product"
-msgstr "Product"
-
-#. module: product_pack
-#: model:ir.model,name:product_pack.model_product_template
-msgid "Product Template"
-msgstr "Product Template"
-
-#. module: product_pack
-#: field:product.pack.line,quantity:0
-#: field:sale.order.line.pack.line,product_uom_qty:0
-msgid "Quantity"
-msgstr "Quantity"
-
-#. module: product_pack
-#: field:product.template,sale_order_pack:0
-msgid "Sale Order Pack"
-msgstr "Sale Order Pack"
-
-#. module: product_pack
-#: help:product.template,sale_order_pack:0
-msgid ""
-"Sale order are packs used on sale orders to calculate a price of a line"
-msgstr "Sale order are packs used on sale orders to calculate a price of a line"
-
-#. module: product_pack
-#: model:ir.model,name:product_pack.model_sale_order
-msgid "Sales Order"
-msgstr "Sales Order"
-
-#. module: product_pack
-#: model:ir.model,name:product_pack.model_sale_order_line
-msgid "Sales Order Line"
-msgstr "Sales Order Line"
-
-#. module: product_pack
-#: view:sale.order.line.pack.line:product_pack.view_sale_line_pack_line_tree
-msgid "Sales Order Line Pack Lines"
-msgstr "Sales Order Line Pack Lines"
-
-#. module: product_pack
-#: help:product.template,pack:0
-msgid "TODO"
-msgstr "TODO"
-
-#. module: product_pack
-#: help:sale.order.line,pack_parent_line_id:0
-msgid "The pack that contains this product."
-msgstr "The pack that contains this product."
-
-#. module: product_pack
-#: field:sale.order.line.pack.line,price_unit:0
-msgid "Unit Price"
-msgstr "Unit Price"
