--- product_pack/views/product_pack_line_views.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record model="ir.ui.view" id="product_pack_line_form">
-        <field name="name">product.pack.line.form</field>
-        <field name="model">product.pack.line</field>
-        <field name="type">form</field>
-        <field name="arch" type="xml">
-            <form string="Pack Product Line">
-                <group>
-                    <field name="product_id"/>
-                    <field name="quantity"/>
-                    <field name="discount" groups="sale.group_discount_per_so_line" invisible="context.get('pack_price_type') == 'fixed_price'"/>
-                </group>
-            </form>
-        </field>
-    </record>
-
-    <record model="ir.ui.view" id="product_pack_line_tree">
-        <field name="name">product.pack.line.tree</field>
-        <field name="model">product.pack.line</field>
-        <field name="type">tree</field>
-        <field name="arch" type="xml">
-            <tree string="Pack Product Lines" editable="bottom">
-                <field name="product_id"/>
-                <field name="quantity"/>
-                <field name="discount" groups="sale.group_discount_per_so_line" invisible="context.get('pack_price_type') == 'fixed_price'"/>
-            </tree>
-        </field>
-    </record>
-</odoo>
