--- product_pack/views/product_template_views.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <!-- Product view modification -->
-    <record model="ir.ui.view" id="product_template_only_form_view">
-        <field name="name">product.template.pack.form</field>
-        <field name="model">product.template</field>
-        <field name="inherit_id" ref="product.product_template_only_form_view" />
-        <field name="arch" type="xml">
-            <field name="type" position="after">
-                <field name="pack"/>
-                <field name="pack_price_type" attrs="{'invisible':[('pack', '=', False)], 'required':[('pack','=',True)]}" context="{'pack_price_type': pack_price_type}"/>
-                <field name="allow_modify_pack" attrs="{'invisible':[('pack', '!=', False), ('pack_price_type', '!=', 'components_price')]}"/>
-            </field>
-            <notebook position="inside">
-                <page string="Pack" attrs="{'invisible': ['|', ('product_variant_count', '>', 1), ('pack', '=', False)]}">
-                    <field name="id" invisible="1"/>
-                    <p attrs="{'invisible': [('id', '!=', False)]}">You must save first to add pack lines</p>
-                    <field name="pack_line_ids" attrs="{'readonly': [('id', '=', False)]}"/>
-                </page>
-            </notebook>
-        </field>
-    </record>
-</odoo>
