--- product_pack/views/sale_order_line_views.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-   <record model="ir.ui.view" id="view_order_line_form2">
-        <field name="name">view.order.line.button</field>
-        <field name="model">sale.order.line</field>
-        <field eval="100" name="priority"/>
-        <field name="arch" type="xml">
-            <form string="Sales Order Lines" create="false">
-                <header>
-                    <field name="state" widget="statusbar" statusbar_visible="draft,confirmed,done" statusbar_colors="{'exception':'red','cancel':'red'}"/>
-                </header>
-                <sheet>
-                    <field name="pack_type" invisible="1"/>
-                    <div attrs="{'invisible':[('pack_type', '!=', 'none_detailed_assited_price')]}">
-                        <div name="product_pack">
-                            <label for="pack_line_ids"/>
-                            <field name="pack_line_ids"/>
-                        </div>
-
-                            <group class="oe_subtotal_footer oe_right" colspan="2" name="pack_total">
-                                <field name="pack_total" widget="monetary" options="{'currency_field': 'currency_id'}"/>
-                                <div class="oe_subtotal_footer_separator" colspan="2"/>
-                                <field name="price_unit" widget="monetary" options="{'currency_field': 'currency_id'}"/>
-                        </group>
-                    </div>
-                </sheet>
-                <footer>
-                    <button class="oe_highlight" name="button_save_data" string="Save" type="object" />
-                    or
-                    <button class="oe_link" special="cancel" string="Cancel" />
-                </footer>
-            </form>
-        </field>
-    </record>
-</odoo>
