+++ sale_product_pack/__init__.py
@@ -1,0 +1,2 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+from . import models
