+++ sale_product_pack/i18n/es.po
@@ -1,0 +1,108 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* sale_product_pack
+#
+# Translators:
+# Juan José Scarafía <scarafia.juanjose@gmail.com>, 2019
+# Gonzalo Ruzafa <gr@ingadhoc.com>, 2019
+# emiquinodoz <eq@ingadhoc.com>, 2019
+# Julián García <jg@adhoc.com.ar>, 2019
+# Valentino Defelice <vd@adhoc.com.ar>, 2019
+# Santiago Crespo <sc@adhoc.com.ar>, 2019
+# Mariana Luz Torres <mt@adhoc.com.ar>, 2019
+# Nicolas Mac Rouillon <nmr@ingadhoc.com>, 2019
+# Katherine Zaoral <katherine.zaoral@gmail.com>, 2019
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2019-09-16 15:45+0000\n"
+"PO-Revision-Date: 2019-09-16 15:45+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: sale_product_pack
+#: model:ir.model.fields,field_description:sale_product_pack.field_sale_order_line__pack_depth
+msgid "Depth"
+msgstr "Profundidad"
+
+#. module: sale_product_pack
+#: model:ir.model.fields,help:sale_product_pack.field_sale_order_line__pack_depth
+msgid "Depth of the product if it is part of a pack."
+msgstr "Profundidad del producto si forma parte de un pack."
+
+#. module: sale_product_pack
+#: model:ir.model.fields,field_description:sale_product_pack.field_sale_order_line__pack_child_line_ids
+msgid "Lines in pack"
+msgstr "Líneas en el pack"
+
+#. module: sale_product_pack
+#: model:ir.model.fields,help:sale_product_pack.field_sale_order_line__pack_type
+msgid "On sale orders or purchase orders:\n"
+"* Detailed - Components Prices: Detail lines with prices.\n"
+"* Detailed - Totalized Price: Detail lines totalicing lines prices on pack (don't show component prices).\n"
+"* Detailed - Fixed Price: Detail lines and use product pack price (ignore line prices).\n"
+"* None Detailed - Totalized Price: Do not detail lines. Do not assist to get pack price using pack lines."
+msgstr "En órdenes de venta u órdenes de compra:\n"
+"* Detallado - Precio de los componentes: Detalla las líneas con precios.\n"
+"* Detallado - Precio totalizado: Detalla las líneas totalizando el precio de las líneas en el pack (no muestro el precio de los componentes).\n"
+"* Detallado - Precio fijo: Detalla las líneas y usa el precio del pack de producto (ignora el precio de las línas).\n"
+"* No detallado - Precio totalizado: No detalla las líneas. No asiste para obtener el precio del pack usando las líneas del pack."
+
+#. module: sale_product_pack
+#: model:ir.model.fields,field_description:sale_product_pack.field_sale_order_line__pack_parent_line_id
+msgid "Pack"
+msgstr "Pack"
+
+#. module: sale_product_pack
+#: model:ir.model.fields,field_description:sale_product_pack.field_sale_order_line__pack_type
+msgid "Pack Type"
+msgstr "Tipo de Pack"
+
+#. module: sale_product_pack
+#: model:ir.model,name:sale_product_pack.model_product_product
+msgid "Product"
+msgstr "Producto"
+
+#. module: sale_product_pack
+#: model:ir.model,name:sale_product_pack.model_product_pack_line
+msgid "Product pack line"
+msgstr "Linea de pack"
+
+#. module: sale_product_pack
+#: model:ir.model,name:sale_product_pack.model_sale_order
+msgid "Sale Order"
+msgstr "Pedido de venta"
+
+#. module: sale_product_pack
+#: model:ir.model.fields,field_description:sale_product_pack.field_product_pack_line__sale_discount
+msgid "Sale discount (%)"
+msgstr "Descuento para ventas"
+
+#. module: sale_product_pack
+#: model:ir.model,name:sale_product_pack.model_sale_order_line
+msgid "Sales Order Line"
+msgstr "Línea de pedido de venta"
+
+#. module: sale_product_pack
+#: model:ir.model.fields,help:sale_product_pack.field_sale_order_line__pack_parent_line_id
+msgid "The pack that contains this product."
+msgstr "El pack que contiene este producto."
+
+#. module: sale_product_pack
+#: code:addons/sale_product_pack/models/sale_order_line.py:76
+#, python-format
+msgid "You can not change this line because is part of a pack included in this order"
+msgstr "No puedes cambiar esta línea porque es parte de un pack incluido en este orden."
+
+#. module: sale_product_pack
+#: code:addons/sale_product_pack/models/sale_order.py:28
+#, python-format
+msgid "You can not delete this line because is part of a pack in this sale order. In order to delete this line you need to delete the pack itself"
+msgstr "No puede eliminar esta línea porque forma parte de un pack en esta orden de venta. Para eliminar esta línea necesitas eliminar el pack."
+
