+++ sale_product_pack/models/__init__.py
@@ -1,0 +1,6 @@
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import product_pack_line
+from . import product_product
+from . import sale_order_line
+from . import sale_order
