+++ sale_product_pack/readme/CONFIGURE.rst
@@ -1,0 +1,6 @@
+To configure this module, you need to:
+
+#. Go to product and set "Product Pack", then set the products has to include
+   in the product pack of the product.
+#. Set type of product pack.
+
