+++ sale_product_pack/readme/CONTRIBUTORS.rst
@@ -1,0 +1,4 @@
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * Ernesto Tejeda
+  * Pedro M. Baeza
