+++ sale_product_pack/readme/DESCRIPTION.rst
@@ -1,0 +1,4 @@
+This module add "Product Pack" functionality to sale orders. For product you
+can add some other product to sell with them, in the sale order you choose
+the product and then automatically added the products that set in Product
+selected.
