+++ sale_product_pack/readme/USAGE.rst
@@ -1,0 +1,3 @@
+To use this module, you need to:
+
+#. Create an sale order and set product that has product pack chosen.
