+++ sale_product_pack/security/ir.model.access.csv
@@ -1,0 +1,2 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_product_pack_line_sale_manager,product.pack.line,model_product_pack_line,sales_team.group_sale_manager,1,1,1,1
